requirements.txt
# Web Framework
fastapi[standard]

# The Parsing Engine (CST)
tree-sitter
tree-sitter-javascript
tree-sitter-cpp
tree-sitter-java
tree-sitter-c-sharp
tree-sitter-python

# AI & LLM Integration
langchain
langchain-google-genai
langgraph
pydantic

# Utilities
python-dotenv
requests